requirements.txt
alembic==1.7.7
Authlib==1.0.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Bootstrap-Flask==2.0.2
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
# greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.32
Werkzeug==2.0.3
wincertstore==0.2
WTForms==3.0.1
